+++ COFantasy.js
@@ -315,6 +315,12 @@ var COFantasy = COFantasy || function() {
     return !inactif;
   }
 
+  function sendChar(charId, msg) {
+    var dest = '';
+    if (charId) dest = 'character|' + charId;
+    sendChat(dest, msg);
+  }
+
   function setState(personnage, etat, value, evt) {
     var token = personnage.token;
     var charId = personnage.charId;
@@ -628,6 +634,11 @@ var COFantasy = COFantasy || function() {
           ab.remove();
         });
         character.remove();
+      });
+    }
+    if (evt.tokens) {
+      evt.tokens.forEach(function(token) {
+        token.remove();
       });
     }
     if (_.has(evt, 'combat')) stateCOF.combat = evt.combat;
@@ -2748,12 +2759,6 @@ var COFantasy = COFantasy || function() {
     attack(playerId, attaquant, targetToken, attackLabel, options);
   }
 
-  function sendChar(charId, msg) {
-    var dest = '';
-    if (charId) dest = 'character|' + charId;
-    sendChat(dest, msg);
-  }
-
   //Si evt est défini, alors on considère qu'il faut y mettre la valeur actuelle
   function updateCurrentBar(token, barNumber, val, evt, maxVal) {
     var prevToken;
@@ -3095,6 +3100,12 @@ var COFantasy = COFantasy || function() {
     setTurnOrder(to, evt);
   }
 
+  function initPerso(personnage, evt, recompute) {
+    initiative([{
+      _id: personnage.token.id
+    }], evt, recompute);
+  }
+
   function setTokenAttr(personnage, attribute, value, evt, msg, maxval) {
     var charId = personnage.charId;
     var token = personnage.token;
@@ -3208,9 +3219,7 @@ var COFantasy = COFantasy || function() {
         var nouveauxPVs = getValeurOfEffet(personnage, 'formeDArbre', niveau * 5);
         updateCurrentBar(token, 1, nouveauxPVs, evt, nouveauxPVs);
         //L'initiative change
-        initiative([{
-          _id: token.id
-        }], evt, true);
+        initPerso(personnage, evt, true);
       }
       return attr;
     }
@@ -6824,6 +6833,10 @@ var COFantasy = COFantasy || function() {
             tempDmg = 0;
           }
         }
+        if (!options.aoe && charAttributeAsBool(target, 'ciblesMultiples')) {
+          dmgTotal = 1;
+          expliquer("La nuée est constituée de très nombreuses cibles, l'attaque ne lui fait qu'1 DM");
+        }
         var pvPerdus = dmgTotal;
         if (target.tempDmg) {
           var oldTempDmg = tempDmg;
@@ -8832,9 +8845,7 @@ var COFantasy = COFantasy || function() {
         if (surveillance(perso)) {
           bonusSurprise += 5;
           setTokenAttr(perso, 'bonusInitEmbuscade', 5, evt, "garde un temps d'avance grâce à son compagnon animal");
-          initiative([{
-            _id: perso.token.id
-          }], evt, true);
+          initPerso(perso, evt, true);
         }
         if (testSurprise !== undefined) {
           testCaracteristique(perso, 'SAG', testSurprise, {
@@ -10150,6 +10161,7 @@ var COFantasy = COFantasy || function() {
   // Ne pas remplacer les inline rolls, il faut les afficher correctement
   function dmgDirects(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 2) {
       error("cof-dmg prend les dégats en argument, avant les options",
@@ -10334,6 +10346,7 @@ var COFantasy = COFantasy || function() {
 
   function interfaceSetState(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 3) {
       error("Pas assez d'arguments pour !cof-set-state", msg.content);
@@ -10417,6 +10430,7 @@ var COFantasy = COFantasy || function() {
 
   function saveState(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 4 ||
       !_.has(cof_states, cmd[1]) || !isCarac(cmd[2])) {
@@ -10769,6 +10783,7 @@ var COFantasy = COFantasy || function() {
 
   function effetTemporaire(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 3) {
       error("Pas assez d'arguments pour !cof-effet-temp", msg.content);
@@ -10965,6 +10980,7 @@ var COFantasy = COFantasy || function() {
 
   function effetCombat(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 2) {
       error("Pas assez d'arguments pour !cof-effet-combat", msg.content);
@@ -11033,7 +11049,7 @@ var COFantasy = COFantasy || function() {
       }
       if (selected.length === 0) return;
       if (limiteRessources(lanceur, options, effet, effet, evt)) return;
-      if (!stateCOF.combat && selected.length > 0) {
+      if (selected.length > 0) {
         initiative(selected, evt);
       }
       iterSelected(selected, function(perso) {
@@ -11075,6 +11091,7 @@ var COFantasy = COFantasy || function() {
 
   function effetIndetermine(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 3) {
       error("Pas assez d'arguments pour !cof-effet", msg.content);
@@ -11707,6 +11724,7 @@ var COFantasy = COFantasy || function() {
       return;
     }
     var options = parseOptions(msg);
+    if (options === undefined) return;
     getSelected(msg, function(selected, playerId) {
       if (selected === undefined || selected.length === 0) {
         sendPlayer(msg, "Pas de cible sélectionnée pour la transe de guérison");
@@ -12087,6 +12105,7 @@ var COFantasy = COFantasy || function() {
 
   function soigner(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd.length < 2) {
       error("Il faut au moins un argument à !cof-soin", cmd);
@@ -12541,6 +12560,7 @@ var COFantasy = COFantasy || function() {
 
   function fortifiant(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 2) {
       error("La fonction !cof-fortifiant attend en argument le rang dans la Voie des élixirs du créateur", cmd);
@@ -12872,6 +12892,7 @@ var COFantasy = COFantasy || function() {
 
   function consommerBaie(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd.length < 2) {
       error("Il faut un argument à !cof-consommer-baie", cmd);
@@ -13199,13 +13220,12 @@ var COFantasy = COFantasy || function() {
         }
       }
     }
-    setTokenAttr({
+    var perso = {
       token: newToken,
       charId: cible.charId
-    }, effet, duree, evt, undefined, getInit());
-    initiative([{
-      _id: newToken.id
-    }], evt);
+    };
+    setTokenAttr(perso, effet, duree, evt, undefined, getInit());
+    initPerso(perso, evt);
   }
 
   //Attention : ne tient pas compte de la rotation !
@@ -13423,6 +13443,7 @@ var COFantasy = COFantasy || function() {
 
   function encaisserUnCoup(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     var evtARefaire = lastEvent();
     if (cmd !== undefined && cmd.length > 1) { //On relance pour un événement particulier
@@ -13739,6 +13760,7 @@ var COFantasy = COFantasy || function() {
 
   function destructionDesMortsVivants(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var args = options.cmd;
     if (args === undefined || args.length < 2) {
       error("Il faut au moins un argument à !cof-destruction-des-morts-vivants", args);
@@ -14364,6 +14386,7 @@ var COFantasy = COFantasy || function() {
   //!cof-creer-elixir token_id nom_token elixir
   function creerElixir(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 3) {
       error("Pas assez d'arguments pour !cof-creer-elixir", msg.content);
@@ -14586,6 +14609,7 @@ var COFantasy = COFantasy || function() {
   //!cof-animer-arbre lanceur-id target-id [rang]
   function animerUnArbre(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 3) {
       error("cof-animer-arbre attend 2 arguments", msg.content);
@@ -14624,9 +14648,7 @@ var COFantasy = COFantasy || function() {
     };
     if (limiteRessources(druide, options, 'animerUnArbre', 'animer un arbre', evt)) return;
     if (!stateCOF.combat) {
-      initiative([{
-        _id: druide.token.id
-      }], evt);
+      initPerso(druide, evt);
     }
     var nomArbre = nouveauNomDePerso('Arbre animé');
     var charArbre = createObj('character', {
@@ -14741,6 +14763,7 @@ var COFantasy = COFantasy || function() {
   //!cof-delivrance @{selected|token_id} @{target|token_id}
   function delivrance(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined || cmd.length < 3) {
       error("cof-delivrance attend 2 arguments", msg.content);
@@ -15139,6 +15162,7 @@ var COFantasy = COFantasy || function() {
       istokenaction: false,
       action: actions
     });
+    return res;
   }
 
   var predateurs = {
@@ -15172,6 +15196,7 @@ var COFantasy = COFantasy || function() {
     loupAlpha: {
       nom: 'Loup alpha',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/59094818/J0yWdxryZFKakJtNGJNNvw/max.jpg?1532612061",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60183959/QAMH6WtyoK78aa4zX_mR_Q/thumb.png?1533898482",
       attributesFiche: {
         NIVEAU: 2,
         FORCE: 16,
@@ -15208,6 +15233,7 @@ var COFantasy = COFantasy || function() {
     worg: {
       nom: 'Grand loup',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/25294798/4dJ_60uP2mw6UJA2elkoXA/max.jpg?1479223790",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60184237/smG5o2-siD2pChhPblO_sQ/thumb.png?1533899118",
       attributesFiche: {
         NIVEAU: 3,
         FORCE: 16,
@@ -15244,6 +15270,7 @@ var COFantasy = COFantasy || function() {
     lion: {
       nom: 'Lion',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/59486104/SngxPIGXDJKdCqsbrXxRYQ/max.jpg?1533041390",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60184437/df1MT2T6lrfo7st02Htxeg/thumb.png?1533899407",
       attributesFiche: {
         NIVEAU: 4,
         FORCE: 20,
@@ -15283,6 +15310,7 @@ var COFantasy = COFantasy || function() {
     grandLion: {
       nom: 'Grand lion',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/59486144/8wHs_5WfEIeL_7dKbALHHA/max.jpg?1533041459",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60186141/mUZzndi9_sYIzdVVNNka_w/thumb.png?1533903070",
       attributesFiche: {
         NIVEAU: 5,
         FORCE: 22,
@@ -15321,6 +15349,7 @@ var COFantasy = COFantasy || function() {
     oursPolaire: {
       nom: 'Ours polaire',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/59486216/UssilagWK_2dfVGuPABBpA/max.png?1533041591",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60186288/B1uAii9G01GcPfQFNozIbw/thumb.png?1533903333",
       attributesFiche: {
         NIVEAU: 6,
         FORCE: 26,
@@ -15353,6 +15382,7 @@ var COFantasy = COFantasy || function() {
     tigreDentsDeSabre: {
       nom: 'Tigre à dents de sabre',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/59486272/f5lUcN3Y9H0thmJPrqa6FQ/max.png?1533041702",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60186469/ShcrgpvgXKiQsLVOyg4SZQ/thumb.png?1533903741",
       attributesFiche: {
         NIVEAU: 7,
         FORCE: 26,
@@ -15392,6 +15422,7 @@ var COFantasy = COFantasy || function() {
     oursPrehistorique: {
       nom: 'Ours préhistorique',
       avatar: "https://s3.amazonaws.com/files.d20.io/images/59486323/V6RVSlBbeRJi_aIaIuGGBw/max.png?1533041814",
+      token: "https://s3.amazonaws.com/files.d20.io/images/60186633/lNHXvCOsvfPMZDQnqJKQVw/thumb.png?1533904189",
       attributesFiche: {
         NIVEAU: 8,
         FORCE: 32,
@@ -15425,22 +15456,37 @@ var COFantasy = COFantasy || function() {
 
   function conjurationPredateur(msg) {
     var options = parseOptions(msg);
+    if (options === undefined) return;
     var cmd = options.cmd;
     if (cmd === undefined) {
       error("Pas de commande", msg.content);
       return;
+    }
+    var renforce = 0;
+    if (cmd.length > 1) {
+      renforce = parseInt(cmd[1]);
+      if (isNaN(renforce)) {
+        error("Il faut un nombre comme premier argument de !cof-conjuration-de-predateur");
+        renforce = 0;
+      }
     }
     getSelected(msg, function(selected, playerId) {
       if (selected === undefined || selected.length === 0) {
         error("pas de lanceur pour la conjuration de prédateurs", msg);
         return;
       }
+      var evt = {
+        type: 'conjuration de prédateurs'
+      };
+      initiative(selected, evt);
       iterSelected(selected, function(invocateur) {
-        var evt = {
-          type: 'conjuration de prédateurs'
-        };
         var pageId = invocateur.token.get('pageid');
         var niveau = ficheAttributeAsInt(invocateur, 'NIVEAU', 1);
+        if (!renforce) {
+          renforce = charAttributeAsInt(invocateur, 'voieDeLaConjuration', 0);
+          if (renforce == 1) renforce = 0;
+        }
+        niveau += renforce;
         var predateur;
         if (niveau < 5) predateur = predateurs.loup;
         else if (niveau < 9) predateur = predateurs.loupAlpha;
@@ -15450,7 +15496,8 @@ var COFantasy = COFantasy || function() {
         else if (niveau < 21) predateur = predateurs.oursPolaire;
         else if (niveau < 23) predateur = predateurs.tigreDentsDeSabre;
         else predateur = predateurs.oursPrehistorique;
-        var nomPredateur = predateur.nom + ' de ' + invocateur.token.get('name');
+        var nomPredateur = 
+          predateur.nom + ' de ' + invocateur.token.get('name');
         var token = createObj('graphic', {
           name: nomPredateur,
           subtype: 'token',
@@ -15467,9 +15514,28 @@ var COFantasy = COFantasy || function() {
           light_angle: 0 //Pour que le joueur ne voit rien par ses yeux
         });
         toFront(token);
-        createCharacter(nomPredateur, playerId, predateur.avatar, token, predateur);
-        addEvent(evt);
+        var charPredateur =
+          createCharacter(nomPredateur, playerId, predateur.avatar, token, predateur);
+        //Tous les prédateurs sont des quadrupèdes
+        createObj('attribute', {
+          name: 'quadrupede',
+          _characterid: charPredateur.id,
+          current: true
+        });
+        //Attribut de predateur conjuré pour la disparition automatique
+        createObj('attribute', {
+          name: 'predateurConjure',
+          _characterid: charPredateur.id,
+          current: 5 + modCarac(invocateur, 'CHARISME'),
+          max: getInit()
+        });
+        evt.characters = [charPredateur];
+        evt.tokens = [token];
+        initiative([{
+          _id: token.id
+        }], evt);
       }); //end iterSelected
+      addEvent(evt);
     }); //end getSelected
   }
 
@@ -16235,7 +16301,13 @@ var COFantasy = COFantasy || function() {
       activation: "continue à agir malgré les blessures",
       actif: "devrait être à terre",
       fin: "subit l'effet de ses blessures"
-    }
+    },
+    predateurConjure: {
+      activation: "apparaît depuis un autre plan",
+      actif: "est un prédateur conjuré",
+      fin: "disparaît",
+      dm: true
+    },
   };
 
   function buildPatternEffets(listeEffets, postfix) {
@@ -16591,6 +16663,7 @@ var COFantasy = COFantasy || function() {
               });
           });
         break;
+      case 'predateurConjure':
       case 'arbreAnime': //effacer le personnage
         //On efface d'abord les attributs et les abilities
         var charAttributes = findObjs({
@@ -16611,20 +16684,27 @@ var COFantasy = COFantasy || function() {
             ab.remove();
           }
         );
-        iterTokensOfAttribute(charId, options.pageId, effet, attrName,
-          function(token) {
-            removeFromTurnTracker(token.id, evt);
-            setToken(token, 'bar1_link', '', evt);
-            setToken(token, 'bar1_value', '', evt);
-            setToken(token, 'bar1_max', '', evt);
-            setToken(token, 'showplayers_bar1', false, evt);
-            setToken(token, 'represents', '', evt);
-            setToken(token, 'showname', false, evt);
-            setToken(token, 'showplayers_name', false, evt);
-            setToken(token, 'name', '', evt);
+        if (effet == 'arbreAnime') {
+          iterTokensOfAttribute(charId, options.pageId, effet, attrName,
+            function(token) {
+              removeFromTurnTracker(token.id, evt);
+              setToken(token, 'bar1_link', '', evt);
+              setToken(token, 'bar1_value', '', evt);
+              setToken(token, 'bar1_max', '', evt);
+              setToken(token, 'showplayers_bar1', false, evt);
+              setToken(token, 'represents', '', evt);
+              setToken(token, 'showname', false, evt);
+              setToken(token, 'showplayers_name', false, evt);
+              setToken(token, 'name', '', evt);
+            });
+        } else if (effet == 'predateurConjure') {
+          iterTokensOfAttribute(charId, options.pageId, effet, attrName, function(token) {
+            token.remove();
           });
+        }
         attr.remove();
-        sendChar(charId, messageEffetTemp[effet].fin);
+        if (options.print) options.print(messageEffetTemp[effet].fin);
+        else sendChar(charId, 'disparaît');
         var arbreChar = getObj('character', charId);
         if (arbreChar) {
           evt.deletedCharacters = evt.deletedCharacters || [];
@@ -16719,13 +16799,13 @@ var COFantasy = COFantasy || function() {
         break;
       default:
     }
-    if (options.attrSave === undefined && !getState({
+    if (options.attrSave === undefined && charId && !getState({
         charId: charId
       }, 'mort')) {
       if (options.print) options.print(messageEffetTemp[effet].fin);
       else sendChar(charId, messageEffetTemp[effet].fin);
     }
-    if (options.gardeAutresAttributs === undefined) {
+    if (options.gardeAutresAttributs === undefined && charId) {
       enleverEffetAttribut(charId, efComplet, attrName, 'Puissant', evt);
       enleverEffetAttribut(charId, efComplet, attrName, 'Valeur', evt);
       enleverEffetAttribut(charId, efComplet, attrName, 'TempeteDeManaIntense', evt);